Clean up webpack files
