+++ config/webpack/envSpecific.js
@@ -1,0 +1,13 @@
+const base = require('./base')
+
+const webpackConfig = (envSpecific) => {
+  const clientConfig = base()
+
+  if (envSpecific) {
+    envSpecific(clientConfig)
+  }
+
+  return clientConfig
+}
+
+module.exports = webpackConfig
